Add SwatchColorPicker base
