+++ packages/office-ui-fabric-react/src/components/SwatchColorPicker/index.ts
@@ -1,2 +1,3 @@
 export * from './SwatchColorPicker';
+export * from './SwatchColorPicker.base';
 export * from './SwatchColorPicker.types';
